Change checksum type to uint16_t

+++ src/encryptor.h
@@ -1,5 +1,6 @@
+#include <stdint.h>
 #include <string.h>
 #include "dynstack.h"
 
-void encryptfile(dynstack *chars, dynstack *fullkey, int checksum, char *path);
-void decryptfile(dynstack *chars, dynstack *fullkey, int checksum, char *path, size_t pathlen);
+void encryptfile(dynstack *chars, dynstack *fullkey, uint16_t checksum, char *path);
+void decryptfile(dynstack *chars, dynstack *fullkey, uint16_t checksum, char *path, size_t pathlen);
